requirements.txt
discord.py
aiohttp
beautifulsoup4
python-dotenv